move seen semaphore to statefulsyncer
