+++ syncer/configuration.go
@@ -65,12 +65,3 @@ func WithAdjustmentWindow(adjustmentWindow int64) Option {
 		s.adjustmentWindow = adjustmentWindow
 	}
 }
-
-// WithSeenConcurrency overrides the number of concurrent
-// invocations of BlockSeen we will make. We default
-// to the value of runtime.NumCPU().
-func WithSeenConcurrency(concurrency int64) Option {
-	return func(s *Syncer) {
-		s.seenSemaphoreSize = concurrency
-	}
-}
